More double to float

+++ src/config.c
@@ -298,7 +298,7 @@ void config_set_thumbstick_offset(float lx, float ly, float rx, float ry) {
     config_cache_synced = false;
 }
 
-void config_set_gyro_offset(double ax, double ay, double az, double bx, double by, double bz) {
+void config_set_gyro_offset(float ax, float ay, float az, float bx, float by, float bz) {
     config_cache.offset_gyro_0_x = ax,
     config_cache.offset_gyro_0_y = ay,
     config_cache.offset_gyro_0_z = az,
@@ -308,7 +308,7 @@ void config_set_gyro_offset(double ax, double ay, double az, double bx, double b
     config_cache_synced = false;
 }
 
-void config_set_accel_offset(double ax, double ay, double az, double bx, double by, double bz) {
+void config_set_accel_offset(float ax, float ay, float az, float bx, float by, float bz) {
     config_cache.offset_accel_0_x = ax,
     config_cache.offset_accel_0_y = ay,
     config_cache.offset_accel_0_z = az,
@@ -318,7 +318,7 @@ void config_set_accel_offset(double ax, double ay, double az, double bx, double 
     config_cache_synced = false;
 }
 
-void config_set_gyro_stddev(double ax, double ay, double az, double bx, double by, double bz) {
+void config_set_gyro_stddev(float ax, float ay, float az, float bx, float by, float bz) {
     config_cache.stddev_gyro_0_x = ax,
     config_cache.stddev_gyro_0_y = ay,
     config_cache.stddev_gyro_0_z = az,
@@ -328,7 +328,7 @@ void config_set_gyro_stddev(double ax, double ay, double az, double bx, double b
     config_cache_synced = false;
 }
 
-void config_set_accel_stddev(double ax, double ay, double az, double bx, double by, double bz) {
+void config_set_accel_stddev(float ax, float ay, float az, float bx, float by, float bz) {
     config_cache.stddev_accel_0_x = ax,
     config_cache.stddev_accel_0_y = ay,
     config_cache.stddev_accel_0_z = az,
@@ -505,7 +505,7 @@ uint8_t config_get_touch_sens_value(uint8_t index) {
     return config_cache.sens_touch_values[index];
 }
 
-double config_get_mouse_sens_value(uint8_t index) {
+float config_get_mouse_sens_value(uint8_t index) {
     return config_cache.sens_mouse_values[index];
 }
 
@@ -521,7 +521,7 @@ void config_set_touch_sens_values(uint8_t* values) {
     config_cache_synced = false;
 }
 
-void config_set_mouse_sens_values(double* values) {
+void config_set_mouse_sens_values(float* values) {
     config_cache.sens_mouse_values[0] = values[0];
     config_cache.sens_mouse_values[1] = values[1];
     config_cache.sens_mouse_values[2] = values[2];
